requirements.txt
asgiref==3.8.1
Django==5.1.4
django-ckeditor==6.7.1
django-ckeditor-5==0.2.15
django-js-asset==2.2.0
pillow==11.0.0
psycopg==3.2.3
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
